+++ src/index.js
@@ -34,6 +34,7 @@ import * as BG from './nationalid/bg/egn.js';
 import * as DK from './nationalid/dk/cpr_number.js';
 import * as FI from './nationalid/fi/henkilotunnus.js';
 import * as NO from './nationalid/no/fodselsnummer.js';
+import * as SE from './nationalid/se/personnummer.js';
 
 const COUNTRY_MODULES = {
   AL: ALB,
@@ -70,6 +71,7 @@ const COUNTRY_MODULES = {
   DK: DK,
   FI: FI,
   NO: NO,
+  SE: SE,
 };
 
 export class NationalID {
@@ -146,6 +148,7 @@ export {
   DK,
   FI,
   NO,
+  SE,
 };
 
 export * from './nationalid/constant.js';